% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_comparison.R
\name{calculate_hd}
\alias{calculate_hd}
\title{Calculate HD (Hamming Distance) between Two DAGs}
\usage{
calculate_hd(bn1, bn2, nodes)
}
\arguments{
\item{bn1}{An object of class \code{bn}, representing the first Bayesian network.}

\item{bn2}{An object of class \code{bn}, representing the second Bayesian network.}

\item{nodes}{A character vector specifying the names of the nodes to include in the comparison.}
}
\value{
A numeric value representing the Hamming Distance (HD) between the two DAGs for the specified nodes.
}
\description{
This function calculates the Hamming Distance (HD) between two directed acyclic graphs (DAGs),
represented as Bayesian networks (\code{bn} objects), for a given set of nodes. The HD measures the
number of differing edges (presence/absence) between two graphs without considering edge orientation.
}
\details{
The function uses the \code{subgraph} function from the \pkg{bnlearn} package to extract the subgraphs
of interest from both networks. It then applies the \code{hamming} function from the \pkg{bnlearn} package
to calculate the edge differences.
}
\examples{
library(bnlearn)

# Create two random Bayesian networks
bn1 <- random.graph(nodes = c("A", "B", "C", "D"))
bn2 <- random.graph(nodes = c("A", "B", "C", "D"))

# Calculate HD for all nodes
hd_all_nodes <- calculate_hd(bn1, bn2, nodes = c("A", "B", "C", "D"))
print(hd_all_nodes)

# Calculate HD for a subset of nodes
hd_subset <- calculate_hd(bn1, bn2, nodes = c("A", "B"))
print(hd_subset)

}
\seealso{
\code{\link[bnlearn]{hamming}}, \code{\link[bnlearn]{subgraph}}
}
