% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_comparison.R
\name{calculate_shd}
\alias{calculate_shd}
\title{Calculate SHD (Structural Hamming Distance) between Two DAGs}
\usage{
calculate_shd(bn1, bn2, nodes)
}
\arguments{
\item{bn1}{An object of class \code{bn}, representing the first Bayesian network.}

\item{bn2}{An object of class \code{bn}, representing the second Bayesian network.}

\item{nodes}{A character vector specifying the names of the nodes to include in the comparison.}
}
\value{
A numeric value representing the Structural Hamming Distance (SHD) between the two DAGs for the specified nodes.
}
\description{
This function calculates the Structural Hamming Distance (SHD) between two directed acyclic graphs (DAGs),
represented as Bayesian networks (\code{bn} objects), for a given set of nodes. The SHD measures the number of
edge differences (additions, deletions, or reversals) needed to transform one graph into the other.
}
\details{
The function uses the \code{subgraph} function from the \pkg{bnlearn} package to extract the subgraphs
of interest from both networks. These subgraphs are converted to \code{graphNEL} objects for compatibility with the
\code{shd} function from the \pkg{pcalg} package.
}
\examples{
library(bnlearn)
library(pcalg)

# Create two random Bayesian networks
bn1 <- random.graph(nodes = c("A", "B", "C", "D"))
bn2 <- random.graph(nodes = c("A", "B", "C", "D"))

# Calculate SHD for all nodes
shd_all_nodes <- calculate_shd(bn1, bn2, nodes = c("A", "B", "C", "D"))
print(shd_all_nodes)

# Calculate SHD for a subset of nodes
shd_subset <- calculate_shd(bn1, bn2, nodes = c("A", "B"))
print(shd_subset)

}
\seealso{
\code{\link[bnlearn]{subgraph}}, \code{\link[pcalg]{shd}}
}
