% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_comparison.R
\name{calculate_tp_fp_fn}
\alias{calculate_tp_fp_fn}
\title{Calculate True Positives (TP), False Positives (FP), and False Negatives (FN) Between Two DAGs}
\usage{
calculate_tp_fp_fn(bn1, bn2, nodes)
}
\arguments{
\item{bn1}{An object of class \code{bn}, representing the first Bayesian network.}

\item{bn2}{An object of class \code{bn}, representing the second Bayesian network.}

\item{nodes}{A character vector specifying the names of the nodes to include in the comparison.}
}
\value{
A list containing the following metrics:
\item{tp}{The number of true positives, representing the edges that exist in both DAGs.}
\item{fp}{The number of false positives, representing the edges that exist in the second DAG but not in the first.}
\item{fn}{The number of false negatives, representing the edges that exist in the first DAG but not in the second.}
}
\description{
This function calculates the number of true positives (TP), false positives (FP), and false negatives (FN)
between two directed acyclic graphs (DAGs), represented as Bayesian networks (\code{bn} objects), for a given set of nodes.
}
\details{
The function uses the \code{subgraph} function from the \pkg{bnlearn} package to extract the subgraphs
for the specified nodes. The \code{compare} function from the \pkg{bnlearn} package is then used to calculate
the metrics based on edge presence and absence.
}
\examples{
library(bnlearn)

# Create two random Bayesian networks
bn1 <- random.graph(nodes = c("A", "B", "C", "D"))
bn2 <- random.graph(nodes = c("A", "B", "C", "D"))

# Calculate TP, FP, FN for all nodes
comparison_metrics <- calculate_tp_fp_fn(bn1, bn2, nodes = c("A", "B", "C", "D"))
print(comparison_metrics)

# Calculate TP, FP, FN for a subset of nodes
subset_metrics <- calculate_tp_fp_fn(bn1, bn2, nodes = c("A", "B"))
print(subset_metrics)

}
\seealso{
\code{\link[bnlearn]{compare}}, \code{\link[bnlearn]{subgraph}}
}
